Changes made:    Removed OpenCV dependency    Other minor changes
