+++ Hardware_Acceleration/Feature_Tutorials/05-using-multiple-cu/reference-files/link.cfg
@@ -1,6 +1,9 @@
 platform=xilinx_u200_gen3x16_xdma_1_202110_1
 debug=1
 
+[profile]
+data=all:all:all:all
+
 [connectivity]
 nk=Filter2DKernel:1
