+++ Hardware_Acceleration/Feature_Tutorials/05-using-multiple-cu/reference-files/link3.cfg
@@ -1,3 +1,4 @@
+platform=xilinx_u200_gen3x16_xdma_1_202110_1
 debug=1
 
 [profile]
