+++ Hardware_Acceleration/Feature_Tutorials/05-using-multiple-cu/reference-files/xrt.ini
@@ -1,3 +1,4 @@
 [Debug] 
 opencl_summary=true
 opencl_trace=true
+data_transfer_trace=accel
